refactor: lottoConstants 상수 객체에 에러 상수 객체 추가 및 파일명 변경
Error.js 을 제거한다.
그에 따라 에러 상수를 constants.js 파일로 옮긴다.
