+++ src/constants/constants.js
@@ -50,3 +50,17 @@ export const FORMAT = Object.freeze({
   NUMBER_MERGER: ', ',
   LOTTO_MERGER: '\n',
 });
+
+export const ERROR = Object.freeze({
+  INVALID_FORMAT: '[ERROR] 잘못된 입력 형식입니다.',
+  CANNOT_CREATE_INSTANCE: '[ERROR] 추상 클래스로 인스턴스를 생성할 수 없습니다.',
+  /**
+   * @param {number} unit - 로또 구입 금액
+   */
+  INVALID_AMOUNT_UNIT: ({ unit }) => `[ERROR] ${unit}원 단위의 금액만 입력해 주세요.`,
+  /**
+   * @param {number} min - 로또 최소 번호
+   * @param {number} max - 로또 최대 번호
+   */
+  INVALID_NUMBER_RANGE: ({ min, max }) => `[ERROR] ${min}이상 ${max}이하의 숫자만 입력해 주세요.`,
+});
